Main layout changes and many classes modified to work with Vector class instead ArrayLists

+++ src/fractureanalysis/plot/PlotVariogramSeries.java
@@ -16,6 +16,7 @@
  */
 package fractureanalysis.plot;
 
+import fractureanalysis.Matrices.Vector;
 import fractureanalysis.analysis.ScanLine;
 import fractureanalysis.analysis.Variograms;
 import java.util.ArrayList;
@@ -42,8 +43,8 @@ public class PlotVariogramSeries {
         ArrayList<Double> variogramValue = new ArrayList();
 
         for (int i = 0; i < distanceList.size(); i++) {      
-            Double value = Variograms.variogram1D(scanline, distanceList.get(i));
-            distance.add(distanceList.get(i));
+            Double value = Variograms.variogram1D(scanline, distanceList.get(i).doubleValue());
+            distance.add(distanceList.get(i).doubleValue());
             variogramValue.add(value);
         }
         return PlotSeries.plotLineSeries(distance, variogramValue);
